+++ app/src/main/java/com/example/radiologi/dokter/home/DataDokterActivity.java
@@ -13,7 +13,7 @@ import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.example.radiologi.DataDokterPagerAdapter;
+import com.example.radiologi.dokter.home.fragments.DataDokterPagerAdapter;
 import com.example.radiologi.LoginActivity;
 import com.example.radiologi.R;
 import com.example.radiologi.SharedPreferenceManager;
