+++ app/src/main/java/com/example/radiologi/dokter/home/fragments/AdapterDokter.java
@@ -1,8 +1,6 @@
-package com.example.radiologi;
+package com.example.radiologi.dokter.home.fragments;
 
 import android.content.Context;
-import android.content.res.ColorStateList;
-import android.graphics.Color;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -15,22 +13,24 @@ import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.radiologi.ListitemDokter;
+import com.example.radiologi.R;
+
 public class AdapterDokter extends RecyclerView.Adapter<AdapterDokter.ViewHolder> {
 
-    private  LayoutInflater inflater;
-    private List<ListitemDokter> listitemDokters;
-    private Context context;
+    private final LayoutInflater inflater;
+    private final List<ListitemDokter> listitemDokters;
 
     private AdapterDokter.OnItemClickListener mListener = null;
 
     public AdapterDokter(Context context) {
         this.listitemDokters = new ArrayList<>();
-        this.context = context;
         this.inflater = LayoutInflater.from(context);
     }
 
+    @NonNull
     @Override
-    public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+    public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View v = inflater.inflate(R.layout.listitem_dokter, parent, false);
         return new ViewHolder(v);
     }
@@ -59,24 +59,14 @@ public class AdapterDokter extends RecyclerView.Adapter<AdapterDokter.ViewHolder
         holder.namaLengkap.setText(listitemDokter.getNamaLengkap());
         holder.jenisKelamin.setText(listitemDokter.getGender());
         holder.tanggalLahir.setText(listitemDokter.getTangLahir());
-        holder.itemView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                mListener.onItemClick(listitemDokter);
-            }
-        });
-        holder.dokterHolder.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                mListener.onItemClick(listitemDokter);
-            }
-        });
+        holder.itemView.setOnClickListener(view -> mListener.onItemClick(listitemDokter));
+        holder.dokterHolder.setOnClickListener(view -> mListener.onItemClick(listitemDokter));
     }
 
     @Override
     public int getItemCount() { return listitemDokters.size(); }
 
-    public class ViewHolder extends RecyclerView.ViewHolder{
+    public static class ViewHolder extends RecyclerView.ViewHolder{
 
         public TextView noRekam;
         public TextView namaLengkap;
@@ -87,8 +77,8 @@ public class AdapterDokter extends RecyclerView.Adapter<AdapterDokter.ViewHolder
         public ViewHolder(View itemView) {
             super(itemView);
 
-            noRekam = (TextView) itemView.findViewById(R.id.noRekam);
-            namaLengkap = (TextView) itemView.findViewById(R.id.namaLengkap);
+            noRekam = itemView.findViewById(R.id.noRekam);
+            namaLengkap = itemView.findViewById(R.id.namaLengkap);
             jenisKelamin = itemView.findViewById(R.id.jenisKelamin);
             tanggalLahir = itemView.findViewById(R.id.tanggalLahir);
             dokterHolder = itemView.findViewById(R.id.dokter);
