+++ app/src/main/res/layout/fragment_data_dokter_baru.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DataDokterBaruFragment">
+    tools:context=".dokter.home.fragments.DataDokterBaruFragment">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
         android:id="@+id/swipe_dokter_data_baru"
