+++ app/src/main/res/layout/fragment_data_dokter_diagnosa.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DataDokterDiagnosaFragment">
+    tools:context=".dokter.home.fragments.DataDokterDiagnosaFragment">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
         android:id="@+id/swipe_dokter_diagnosa"
